configure.ac: fix complaint by automake of missing AC_PROG_YACC
If automake sees you have a Yacc source but YACC is undefined
it will complain.

Makefile.am: error: Yacc source seen but 'YACC' is undefined
Makefile.am:   The usual way to define 'YACC' is to add 'AC_PROG_YACC'
Makefile.am:   to 'configure.ac' and run 'autoconf' again.
autoreconf: automake failed with exit status:

The recommended way to resolve this is to use AC_PROG_YACC on configure.ac,
but that is only if we care for yacc. Since we clearly are not fans of
it and seem to prefer bison, add the missing m4 helper which defines
AX_PROG_BISON().

I've taken this from the mesa project [0].

[0] https://gitlab.collabora.com/spurv/mesa/-/blob/collabora_master/m4/ax_prog_bison.m4

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>